[OpenVINO] Fix mismatch operand type in MoE for Trinity (Afmoe)

+++ optimum/exporters/openvino/model_patcher.py
@@ -7469,14 +7469,14 @@ class AfmoeModelPatcher(ModelPatcher):
                 afmoe_moe.down_projs = torch.concat(
                     tuple(afmoe_moe.experts[i].down_proj.weight.unsqueeze(0) for i in range(num_experts)),
                     dim=0,
-                ).transpose(1, 2)
+                ).transpose(1, 2).float()
                 afmoe_moe.gate_projs = torch.concat(
                     tuple(afmoe_moe.experts[i].gate_proj.weight.unsqueeze(0) for i in range(num_experts)),
                     dim=0,
-                ).transpose(1, 2)
+                ).transpose(1, 2).float()
                 afmoe_moe.up_projs = torch.concat(
                     tuple(afmoe_moe.experts[i].up_proj.weight.unsqueeze(0) for i in range(num_experts)), dim=0
-                ).transpose(1, 2)
+                ).transpose(1, 2).float()
 
     def __exit__(self, exc_type, exc_value, traceback):
         super().__exit__(exc_type, exc_value, traceback)
